KIWI-BAV-1336 Refactored step def files
