+++ test/browser/step_definitions/BAVLandingPageSteps.step.js
@@ -1,14 +1,9 @@
 const { When, Then } = require("@cucumber/cucumber");
-const { BavLandingPage, AccountDetailsPage } = require("../pages");
+const { BavLandingPage, AccountDetailsPage, AbortPage } = require("../pages");
 
 When("the user clicks on Continue button", async function () {
   const landingPage = new BavLandingPage(await this.page);
   await landingPage.continueToAccountDetails();
-});
-
-Then("the user is directed to the Account Details screen", async function () {
-  const accountDetailsPage = new AccountDetailsPage(await this.page);
-  await accountDetailsPage.isCurrentPage();
 });
 
 When(
@@ -18,6 +13,11 @@ When(
     await landingPage.clickHowWeUseBankDetails();
   }
 );
+
+Then("they are routed to the BAV Landing Page", async function () {
+  const landingPage = new BavLandingPage(await this.page);
+  expect(await landingPage.isCurrentPage()).toBeTruthy();
+});
 
 Then(
   "the 'How we use your bank details' information is presented to the user",
